requirements.txt
numpy==1.23.5
pandas==1.5.3
tabpfn==2.0.9
pycox==0.3.0
scikit-survival==0.24.1
torchtuples==0.2.2
lifelines==0.24.3
scikit-learn==1.5.2
scipy==1.15.3
matplotlib==3.7.1
seaborn==0.12.2